requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
rapidfuzz>=3.5.0
sentence-transformers>=2.2.0
torch>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
unidecode>=1.3.0
